+++ Sources/BraintreeShopperInsights/BTButtonType.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+/// The button type to be displayed or presented
+/// Warning: This module is in beta. It's public API may change or be removed in future releases.
+
+public enum BTButtonType: String {
+    case payPal = "Paypal"
+    case venmo = "Venmo"
+    case other = "Other"
+}